Run core evaluator against canonical fixtures
