+++ products/feature-flagging/feature-flagging-core/src/test/java/datadog/openfeature/internal/core/FlagEvaluatorTest.java
@@ -3,8 +3,12 @@ package datadog.openfeature.internal.core;
 import static java.util.Collections.emptyList;
 import static java.util.Collections.emptyMap;
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertNull;
+import static org.junit.jupiter.api.Assertions.assertFalse;
 
+import com.squareup.moshi.JsonAdapter;
+import com.squareup.moshi.JsonDataException;
+import com.squareup.moshi.Moshi;
+import com.squareup.moshi.Types;
 import datadog.openfeature.internal.core.EvaluationResult.Error;
 import datadog.openfeature.internal.core.EvaluationResult.Reason;
 import datadog.openfeature.internal.core.FlagEvaluator.ValueKind;
@@ -15,24 +19,64 @@ import datadog.trace.api.featureflag.ufc.v1.Environment;
 import datadog.trace.api.featureflag.ufc.v1.Flag;
 import datadog.trace.api.featureflag.ufc.v1.Rule;
 import datadog.trace.api.featureflag.ufc.v1.ServerConfiguration;
-import datadog.trace.api.featureflag.ufc.v1.Shard;
-import datadog.trace.api.featureflag.ufc.v1.ShardRange;
 import datadog.trace.api.featureflag.ufc.v1.Split;
 import datadog.trace.api.featureflag.ufc.v1.ValueType;
 import datadog.trace.api.featureflag.ufc.v1.Variant;
+import java.io.IOException;
+import java.lang.reflect.Type;
+import java.nio.charset.StandardCharsets;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.Paths;
 import java.util.ArrayList;
-import java.util.Date;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.MethodSource;
 
 class FlagEvaluatorTest {
 
+  private static final String CANONICAL_FIXTURE_PATH =
+      "dd-smoke-tests/openfeature/src/test/resources/ffe-system-test-data";
+  private static final Type FIXTURE_LIST_TYPE =
+      Types.newParameterizedType(List.class, FixtureCase.class);
+  private static final JsonAdapter<List<FixtureCase>> FIXTURE_LIST_ADAPTER =
+      new Moshi.Builder().build().adapter(FIXTURE_LIST_TYPE);
+  private static final UfcParser UFC_PARSER = new UfcParser();
+
   private final FlagEvaluator evaluator = new FlagEvaluator();
 
+  @ParameterizedTest(name = "{0}")
+  @MethodSource("canonicalTestCases")
+  void evaluatesCanonicalFixture(final FixtureCase testCase) throws IOException {
+    final ValueKind valueKind = valueKind(testCase.variationType);
+    final Object defaultValue = FlagEvaluator.mapValue(valueKind, testCase.defaultValue);
+    final EvaluationResult result =
+        evaluator.evaluate(
+            loadCanonicalConfiguration(),
+            valueKind,
+            testCase.flag,
+            defaultValue,
+            context(testCase));
+
+    assertEquals(FlagEvaluator.mapValue(valueKind, testCase.result.value), result.value);
+    assertEquals(Reason.valueOf(testCase.result.reason), result.reason);
+    if (testCase.result.errorCode != null) {
+      assertEquals(Error.valueOf(testCase.result.errorCode), result.error);
+    }
+  }
+
   @Test
-  void evaluatesStaticValuesAndMetadata() {
+  void canonicalFixturesArePresent() throws IOException {
+    assertFalse(canonicalTestCases().isEmpty());
+  }
+
+  @Test
+  void evaluatesMetadataExcludedFromCanonicalFixtures() {
     final EvaluationResult result =
         evaluate(staticFlag(ValueType.STRING, "hello"), ValueKind.STRING, context("subject"));
 
@@ -45,41 +89,13 @@ class FlagEvaluatorTest {
   }
 
   @Test
-  void mapsAllSupportedValueKinds() {
-    assertEquals(
-        true, evaluate(staticFlag(ValueType.BOOLEAN, 1), ValueKind.BOOLEAN, context("id")).value);
-    assertEquals(
-        42,
-        evaluate(staticFlag(ValueType.INTEGER, "42.9"), ValueKind.INTEGER, context("id")).value);
-    assertEquals(
-        42D, evaluate(staticFlag(ValueType.NUMERIC, "42"), ValueKind.DOUBLE, context("id")).value);
-    assertEquals(
-        "42", evaluate(staticFlag(ValueType.STRING, 42), ValueKind.STRING, context("id")).value);
-    assertEquals(
-        map("nested", true),
-        evaluate(staticFlag(ValueType.JSON, map("nested", true)), ValueKind.OBJECT, context("id"))
-            .value);
-    assertNull(FlagEvaluator.mapValue(ValueKind.STRING, null));
-  }
-
-  @Test
-  void reportsInputAndConfigurationErrors() {
+  void reportsErrorsOutsideCanonicalFixtureContract() {
     assertError(null, ValueKind.STRING, context("id"), Error.PROVIDER_NOT_READY);
     assertError(
         snapshot(staticFlag(ValueType.STRING, "value")),
         ValueKind.STRING,
         null,
         Error.INVALID_CONTEXT);
-
-    final EvaluationResult missing =
-        evaluator.evaluate(
-            snapshot(staticFlag(ValueType.STRING, "value")),
-            ValueKind.STRING,
-            "missing",
-            "default",
-            context("id"));
-    assertEquals(Error.FLAG_NOT_FOUND, missing.error);
-
     assertError(
         snapshot(
             new Flag("flag", true, ValueType.STRING, map("on", new Variant("on", "value")), null)),
@@ -93,10 +109,7 @@ class FlagEvaluatorTest {
         Error.TYPE_MISMATCH);
     assertError(
         snapshot(
-            flag(
-                ValueType.STRING,
-                map("on", new Variant("on", "value")),
-                list(split("missing", emptyList())))),
+            flag(ValueType.STRING, map("on", new Variant("on", "value")), list(split("missing")))),
         ValueKind.STRING,
         context("id"),
         Error.GENERAL);
@@ -108,88 +121,7 @@ class FlagEvaluatorTest {
   }
 
   @Test
-  void returnsDisabledAndDefaultResults() {
-    final Flag disabled =
-        new Flag(
-            "flag", false, ValueType.STRING, map("on", new Variant("on", "value")), emptyList());
-    assertEquals(Reason.DISABLED, evaluate(disabled, ValueKind.STRING, context("id")).reason);
-
-    final Flag noSplits =
-        flag(ValueType.STRING, map("on", new Variant("on", "value")), emptyList());
-    assertEquals(Reason.DEFAULT, evaluate(noSplits, ValueKind.STRING, context("id")).reason);
-
-    final long now = System.currentTimeMillis();
-    final Flag inactive =
-        new Flag(
-            "flag",
-            true,
-            ValueType.STRING,
-            map("on", new Variant("on", "value")),
-            list(
-                new Allocation(
-                    "future",
-                    emptyList(),
-                    new Date(now + 60_000),
-                    null,
-                    list(split("on", emptyList())),
-                    false),
-                new Allocation(
-                    "past",
-                    emptyList(),
-                    null,
-                    new Date(now - 60_000),
-                    list(split("on", emptyList())),
-                    false)));
-    assertEquals(Reason.DEFAULT, evaluate(inactive, ValueKind.STRING, context("id")).reason);
-  }
-
-  @Test
-  void evaluatesAllRuleOperators() {
-    final Rule matchingRule =
-        new Rule(
-            list(
-                condition(ConditionOperator.MATCHES, "country", "^U"),
-                condition(ConditionOperator.NOT_MATCHES, "country", "^C"),
-                condition(ConditionOperator.ONE_OF, "tier", list("free", "paid")),
-                condition(ConditionOperator.NOT_ONE_OF, "tier", list("blocked")),
-                condition(ConditionOperator.GTE, "age", 18),
-                condition(ConditionOperator.GT, "age", 17),
-                condition(ConditionOperator.LTE, "age", 18),
-                condition(ConditionOperator.LT, "age", 19),
-                condition(ConditionOperator.IS_NULL, "missing", true),
-                condition(ConditionOperator.IS_NULL, "country", false),
-                condition(ConditionOperator.ONE_OF, "score", list("18"))));
-    final Flag flag =
-        new Flag(
-            "flag",
-            true,
-            ValueType.STRING,
-            map("on", new Variant("on", "matched")),
-            list(
-                new Allocation(
-                    "targeted",
-                    list(new Rule(emptyList()), matchingRule),
-                    null,
-                    null,
-                    list(split("on", emptyList())),
-                    false)));
-    final EvaluationContext matching =
-        new EvaluationContext(
-            "subject", map("country", "US", "tier", "paid", "age", 18, "score", 18));
-
-    assertEquals(Reason.TARGETING_MATCH, evaluate(flag, ValueKind.STRING, matching).reason);
-    assertEquals(
-        Reason.DEFAULT,
-        evaluate(
-                flag,
-                ValueKind.STRING,
-                new EvaluationContext(
-                    "subject", map("country", "CA", "tier", "paid", "age", 18, "score", 18)))
-            .reason);
-  }
-
-  @Test
-  void mapsInvalidRulesToEvaluationErrors() {
+  void reportsMalformedRuleErrorsOutsideCanonicalFixtureContract() {
     final Flag invalidRegex = targetedFlag(condition(ConditionOperator.MATCHES, "value", "["));
     assertEquals(
         Error.PARSE_ERROR,
@@ -203,40 +135,74 @@ class FlagEvaluatorTest {
             .error);
   }
 
-  @Test
-  void evaluatesShardsAndRequiresTargetingKey() {
-    final Shard matchingShard = new Shard("salt", list(new ShardRange(0, 1)), 1);
-    final Flag sharded =
-        flag(
-            ValueType.STRING,
-            map("on", new Variant("on", "value")),
-            list(split("on", list(matchingShard))));
-
-    assertEquals(Reason.SPLIT, evaluate(sharded, ValueKind.STRING, context("id")).reason);
-    assertEquals(
-        Error.TARGETING_KEY_MISSING, evaluate(sharded, ValueKind.STRING, context(null)).error);
-
-    final Shard invalidShard = new Shard("salt", emptyList(), 0);
-    final Flag unmatched =
-        flag(
-            ValueType.STRING,
-            map("on", new Variant("on", "value")),
-            list(split("on", list(invalidShard))));
-    assertEquals(Reason.DEFAULT, evaluate(unmatched, ValueKind.STRING, context("id")).reason);
+  private static ServerConfiguration loadCanonicalConfiguration() throws IOException {
+    return UFC_PARSER.parse(Files.readAllBytes(fixtureRoot().resolve("ufc-config.json")));
   }
 
-  @Test
-  void contextUsesTargetingKeyAsDefaultIdAndCopiesAttributes() {
-    final java.util.Map<String, Object> mutable = new java.util.LinkedHashMap<>();
-    mutable.put("name", "value");
-    final EvaluationContext context = new EvaluationContext("subject", mutable);
-    mutable.put("name", "changed");
+  private static List<FixtureCase> canonicalTestCases() throws IOException {
+    final Path evaluationCases = fixtureRoot().resolve("evaluation-cases");
+    final List<FixtureCase> result = new ArrayList<>();
 
-    assertEquals("subject", context.attribute("id"));
-    assertEquals("value", context.attribute("name"));
-    assertEquals(
-        "explicit", new EvaluationContext("subject", map("id", "explicit")).attribute("id"));
-    assertNull(new EvaluationContext("subject", null).attribute("missing"));
+    try (final Stream<Path> paths = Files.list(evaluationCases)) {
+      final List<Path> files =
+          paths
+              .filter(path -> path.getFileName().toString().endsWith(".json"))
+              .sorted((left, right) -> left.getFileName().compareTo(right.getFileName()))
+              .collect(Collectors.toList());
+      for (final Path file : files) {
+        final List<FixtureCase> testCases = FIXTURE_LIST_ADAPTER.fromJson(read(file));
+        if (testCases == null) {
+          throw new JsonDataException("Fixture file did not contain an array: " + file);
+        }
+        for (int index = 0; index < testCases.size(); index++) {
+          final FixtureCase testCase = testCases.get(index);
+          testCase.fileName = file.getFileName().toString();
+          testCase.index = index;
+          result.add(testCase);
+        }
+      }
+    }
+
+    return result;
+  }
+
+  private static Path fixtureRoot() {
+    Path directory = Paths.get("").toAbsolutePath();
+    while (directory != null) {
+      final Path candidate = directory.resolve(CANONICAL_FIXTURE_PATH);
+      if (Files.exists(candidate.resolve("ufc-config.json"))
+          && Files.isDirectory(candidate.resolve("evaluation-cases"))) {
+        return candidate;
+      }
+      directory = directory.getParent();
+    }
+    throw new IllegalStateException("Unable to find canonical FFE fixtures");
+  }
+
+  private static String read(final Path path) throws IOException {
+    return new String(Files.readAllBytes(path), StandardCharsets.UTF_8);
+  }
+
+  private static ValueKind valueKind(final String variationType) {
+    switch (variationType) {
+      case "BOOLEAN":
+        return ValueKind.BOOLEAN;
+      case "INTEGER":
+        return ValueKind.INTEGER;
+      case "NUMERIC":
+        return ValueKind.DOUBLE;
+      case "STRING":
+        return ValueKind.STRING;
+      case "JSON":
+        return ValueKind.OBJECT;
+      default:
+        throw new IllegalArgumentException("Unsupported variationType: " + variationType);
+    }
+  }
+
+  private static EvaluationContext context(final FixtureCase testCase) {
+    return new EvaluationContext(
+        testCase.targetingKey, testCase.attributes == null ? emptyMap() : testCase.attributes);
   }
 
   private EvaluationResult evaluate(
@@ -261,7 +227,7 @@ class FlagEvaluatorTest {
   }
 
   private static Flag staticFlag(final ValueType type, final Object value) {
-    return flag(type, map("on", new Variant("on", value)), list(split("on", emptyList())));
+    return flag(type, map("on", new Variant("on", value)), list(split("on")));
   }
 
   private static Flag targetedFlag(final ConditionConfiguration condition) {
@@ -276,7 +242,7 @@ class FlagEvaluatorTest {
                 list(new Rule(list(condition))),
                 null,
                 null,
-                list(split("on", emptyList())),
+                list(split("on")),
                 false)));
   }
 
@@ -290,8 +256,8 @@ class FlagEvaluatorTest {
         list(new Allocation("allocation", emptyList(), null, null, splits, true)));
   }
 
-  private static Split split(final String variationKey, final List<Shard> shards) {
-    return new Split(shards, variationKey, emptyMap(), 7);
+  private static Split split(final String variationKey) {
+    return new Split(emptyList(), variationKey, emptyMap(), 7);
   }
 
   private static ConditionConfiguration condition(
@@ -315,5 +281,27 @@ class FlagEvaluatorTest {
       result.put((K) keyValues[index], (V) keyValues[index + 1]);
     }
     return result;
+  }
+
+  private static final class FixtureCase {
+    Map<String, Object> attributes = emptyMap();
+    Object defaultValue;
+    String flag;
+    FixtureResult result;
+    String targetingKey;
+    String variationType;
+    transient String fileName;
+    transient int index;
+
+    @Override
+    public String toString() {
+      return fileName + "[" + index + "] flag=" + flag;
+    }
+  }
+
+  private static final class FixtureResult {
+    Object value;
+    String reason;
+    String errorCode;
   }
 }
